update: 调整max_flow_ctl类型为int,以便于配置

+++ config/config.go
@@ -420,7 +420,7 @@ type Osc struct {
 	OscMaxLag int `toml:"osc_max_lag" json:"osc_max_lag"`
 
 	// 类似--max-lag，检查集群暂停流量控制所花费的平均时间（仅适用于PXC 5.6及以上版本）
-	OscMaxFlowCtl float32 `toml:"osc_max_flow_ctl" json:"osc_max_flow_ctl"`
+	OscMaxFlowCtl int `toml:"osc_max_flow_ctl" json:"osc_max_flow_ctl"`
 
 	// 对应参数pt-online-schema-change中的参数--[no]check-alter。默认值：ON
 	OscCheckAlter bool `toml:"osc_check_alter" json:"osc_check_alter"`
